requirements.txt
django==3.2
djangorestframework==3.11.0
pytz==2019.3
psycopg2-binary==2.8.4
requests==2.23.0
django-mptt==0.12.0
